requirements.txt
opencv-python~=4.5.2.54
PyYAML~=5.4.1
Flask~=2.0.2
numpy~=1.19.4
opcua~=0.98.13
requests~=2.26.0
setuptools~=59.2.0